# isp-framework/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DotMac ISP Framework",
    "description": "Comprehensive ISP management system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/services/plans": {
      "get": {
        "tags": [
          "services",
          "services"
        ],
        "summary": "List Service Plans",
        "description": "List all service plans with optional filtering.",
        "operationId": "list_service_plans_api_v1_services_plans_get",
        "parameters": [
          {
            "name": "service_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ServiceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Type"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "is_public",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Public"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicePlanResponse"
                  },
                  "title": "Response List Service Plans Api V1 Services Plans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "services",
          "services"
        ],
        "summary": "Create Service Plan",
        "description": "Create a new service plan.",
        "operationId": "create_service_plan_api_v1_services_plans_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicePlanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicePlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/services/plans/{plan_id}": {
      "get": {
        "tags": [
          "services",
          "services"
        ],
        "summary": "Get Service Plan",
        "description": "Get a specific service plan by ID.",
        "operationId": "get_service_plan_api_v1_services_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServicePlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/services/instances": {
      "get": {
        "tags": [
          "services",
          "services"
        ],
        "summary": "List Service Instances",
        "description": "List all service instances with optional filtering.",
        "operationId": "list_service_instances_api_v1_services_instances_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ServiceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceInstanceResponse"
                  },
                  "title": "Response List Service Instances Api V1 Services Instances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/services/instances/{instance_id}/modify": {
      "put": {
        "tags": [
          "services",
          "services"
        ],
        "summary": "Modify Service",
        "description": "Modify an existing service instance.",
        "operationId": "modify_service_api_v1_services_instances__instance_id__modify_put",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Instance Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceModificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceInstanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BandwidthUnit": {
        "type": "string",
        "enum": [
          "kbps",
          "mbps",
          "gbps"
        ],
        "title": "BandwidthUnit",
        "description": "Bandwidth unit enumeration."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServiceModificationRequest": {
        "properties": {
          "new_service_plan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Service Plan Id"
          },
          "add_addons": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Add Addons"
          },
          "remove_addons": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Remove Addons"
          },
          "change_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Address"
          },
          "modification_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Modification Notes"
          },
          "effective_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          }
        },
        "type": "object",
        "title": "ServiceModificationRequest",
        "description": "Schema for service modification requests."
      },
...
      "ServicePlanCreate": {
        "properties": {
          "plan_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Plan Code"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "service_type": {
            "$ref": "#/components/schemas/ServiceType"
          },
          "monthly_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Monthly Price"
          },
          "setup_fee": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Setup Fee",
            "default": 0
          },
          "cancellation_fee": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Cancellation Fee",
            "default": 0
          },
          "download_speed": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Speed"
          },
          "upload_speed": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Speed"
          },
          "bandwidth_unit": {
            "$ref": "#/components/schemas/BandwidthUnit",
            "default": "mbps"
          },
          "data_allowance": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Allowance"
          },
          "features": {
            "additionalProperties": true,
            "type": "object",
            "title": "Features"
          },
          "technical_specs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Technical Specs"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": true
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          },
          "min_contract_months": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min Contract Months",
            "default": 0
          },
          "max_contract_months": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Contract Months"
          }
        },
        "type": "object",
        "required": [
          "plan_code",
          "name",
          "service_type",
          "monthly_price"
        ],
        "title": "ServicePlanCreate",
        "description": "Schema for creating service plans."
      },
      "ServicePlanResponse": {
        "properties": {
          "plan_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Plan Code"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "service_type": {
            "$ref": "#/components/schemas/ServiceType"
          },
          "monthly_price": {
            "type": "string",
            "title": "Monthly Price"
          },
          "setup_fee": {
            "type": "string",
            "title": "Setup Fee",
            "default": 0
          },
          "cancellation_fee": {
            "type": "string",
            "title": "Cancellation Fee",
            "default": 0
          },
          "download_speed": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Speed"
          },
          "upload_speed": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Speed"
          },
          "bandwidth_unit": {
            "$ref": "#/components/schemas/BandwidthUnit",
            "default": "mbps"
          },
          "data_allowance": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Allowance"
          },
          "features": {
            "additionalProperties": true,
            "type": "object",
            "title": "Features"
          },
          "technical_specs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Technical Specs"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": true
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          },
          "min_contract_months": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min Contract Months",
            "default": 0
          },
          "max_contract_months": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Contract Months"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted",
            "default": false
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "plan_code",
          "name",
          "service_type",
          "monthly_price",
          "id",
          "created_at",
          "updated_at",
          "tenant_id"
        ],
        "title": "ServicePlanResponse",
        "description": "Schema for service plan responses."
      },
      "ServiceStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "pending",
          "suspended",
          "cancelled",
          "maintenance"
        ],
        "title": "ServiceStatus",
        "description": "Service status enumeration."
      },
      "ServiceType": {
        "type": "string",
        "enum": [
          "internet",
          "phone",
          "tv",
          "bundle",
          "hosting",
          "cloud",
          "managed_services"
        ],
        "title": "ServiceType",
        "description": "Service type enumeration."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}